+++ avx2/polyvec.c
@@ -27,17 +27,17 @@ void PQCLEAN_NAMESPACE_polyvec_compress(uint8_t * restrict r, polyvec * restrict
       for(k=0;k<8;k++)
         t[k] = ((((uint32_t)a->vec[i].coeffs[8*j+k] << 11) + KYBER_Q/2) / KYBER_Q) & 0x7ff;
 
-      r[11*j+ 0] =  t[0] & 0xff;
-      r[11*j+ 1] = (t[0] >>  8) | ((t[1] & 0x1f) << 3);
-      r[11*j+ 2] = (t[1] >>  5) | ((t[2] & 0x03) << 6);
-      r[11*j+ 3] = (t[2] >>  2) & 0xff;
-      r[11*j+ 4] = (t[2] >> 10) | ((t[3] & 0x7f) << 1);
-      r[11*j+ 5] = (t[3] >>  7) | ((t[4] & 0x0f) << 4);
-      r[11*j+ 6] = (t[4] >>  4) | ((t[5] & 0x01) << 7);
-      r[11*j+ 7] = (t[5] >>  1) & 0xff;
-      r[11*j+ 8] = (t[5] >>  9) | ((t[6] & 0x3f) << 2);
-      r[11*j+ 9] = (t[6] >>  6) | ((t[7] & 0x07) << 5);
-      r[11*j+10] = (t[7] >>  3);
+      r[11*j+ 0] = (uint8_t)( t[0] & 0xff);
+      r[11*j+ 1] = (uint8_t)((t[0] >>  8) | ((t[1] & 0x1f) << 3));
+      r[11*j+ 2] = (uint8_t)((t[1] >>  5) | ((t[2] & 0x03) << 6));
+      r[11*j+ 3] = (uint8_t)((t[2] >>  2) & 0xff);
+      r[11*j+ 4] = (uint8_t)((t[2] >> 10) | ((t[3] & 0x7f) << 1));
+      r[11*j+ 5] = (uint8_t)((t[3] >>  7) | ((t[4] & 0x0f) << 4));
+      r[11*j+ 6] = (uint8_t)((t[4] >>  4) | ((t[5] & 0x01) << 7));
+      r[11*j+ 7] = (uint8_t)((t[5] >>  1) & 0xff);
+      r[11*j+ 8] = (uint8_t)((t[5] >>  9) | ((t[6] & 0x3f) << 2));
+      r[11*j+ 9] = (uint8_t)((t[6] >>  6) | ((t[7] & 0x07) << 5));
+      r[11*j+10] = (uint8_t)((t[7] >>  3));
     }
     r += 352;
   }
@@ -50,11 +50,11 @@ void PQCLEAN_NAMESPACE_polyvec_compress(uint8_t * restrict r, polyvec * restrict
       for(k=0;k<4;k++)
         t[k] = ((((uint32_t)a->vec[i].coeffs[4*j+k] << 10) + KYBER_Q/2) / KYBER_Q) & 0x3ff;
 
-      r[5*j+ 0] =  t[0] & 0xff;
-      r[5*j+ 1] = (t[0] >>  8) | ((t[1] & 0x3f) << 2);
-      r[5*j+ 2] = (t[1] >>  6) | ((t[2] & 0x0f) << 4);
-      r[5*j+ 3] = (t[2] >>  4) | ((t[3] & 0x03) << 6);
-      r[5*j+ 4] = (t[3] >>  2);
+      r[5*j+ 0] = (uint8_t)( t[0] & 0xff);
+      r[5*j+ 1] = (uint8_t)((t[0] >>  8) | ((t[1] & 0x3f) << 2));
+      r[5*j+ 2] = (uint8_t)((t[1] >>  6) | ((t[2] & 0x0f) << 4));
+      r[5*j+ 3] = (uint8_t)((t[2] >>  4) | ((t[3] & 0x03) << 6));
+      r[5*j+ 4] = (uint8_t)((t[3] >>  2));
     }
     r += 320;
   }
@@ -80,14 +80,14 @@ void PQCLEAN_NAMESPACE_polyvec_decompress(polyvec * restrict r, const uint8_t * 
   {
     for(j=0;j<KYBER_N/8;j++)
     {
-      r->vec[i].coeffs[8*j+0] =  (((a[11*j+ 0]       | (((uint32_t)a[11*j+ 1] & 0x07) << 8)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+1] = ((((a[11*j+ 1] >> 3) | (((uint32_t)a[11*j+ 2] & 0x3f) << 5)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+2] = ((((a[11*j+ 2] >> 6) | (((uint32_t)a[11*j+ 3] & 0xff) << 2) | (((uint32_t)a[11*j+ 4] & 0x01) << 10)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+3] = ((((a[11*j+ 4] >> 1) | (((uint32_t)a[11*j+ 5] & 0x0f) << 7)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+4] = ((((a[11*j+ 5] >> 4) | (((uint32_t)a[11*j+ 6] & 0x7f) << 4)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+5] = ((((a[11*j+ 6] >> 7) | (((uint32_t)a[11*j+ 7] & 0xff) << 1) | (((uint32_t)a[11*j+ 8] & 0x03) <<  9)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+6] = ((((a[11*j+ 8] >> 2) | (((uint32_t)a[11*j+ 9] & 0x1f) << 6)) * KYBER_Q) + 1024) >> 11;
-      r->vec[i].coeffs[8*j+7] = ((((a[11*j+ 9] >> 5) | (((uint32_t)a[11*j+10] & 0xff) << 3)) * KYBER_Q) + 1024) >> 11;
+      r->vec[i].coeffs[8*j+0] = (int16_t)( (((a[11*j+ 0]       | (((uint32_t)a[11*j+ 1] & 0x07) << 8)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+1] = (int16_t)(((((a[11*j+ 1] >> 3) | (((uint32_t)a[11*j+ 2] & 0x3f) << 5)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+2] = (int16_t)(((((a[11*j+ 2] >> 6) | (((uint32_t)a[11*j+ 3] & 0xff) << 2) | (((uint32_t)a[11*j+ 4] & 0x01) << 10)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+3] = (int16_t)(((((a[11*j+ 4] >> 1) | (((uint32_t)a[11*j+ 5] & 0x0f) << 7)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+4] = (int16_t)(((((a[11*j+ 5] >> 4) | (((uint32_t)a[11*j+ 6] & 0x7f) << 4)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+5] = (int16_t)(((((a[11*j+ 6] >> 7) | (((uint32_t)a[11*j+ 7] & 0xff) << 1) | (((uint32_t)a[11*j+ 8] & 0x03) <<  9)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+6] = (int16_t)(((((a[11*j+ 8] >> 2) | (((uint32_t)a[11*j+ 9] & 0x1f) << 6)) * KYBER_Q) + 1024) >> 11);
+      r->vec[i].coeffs[8*j+7] = (int16_t)(((((a[11*j+ 9] >> 5) | (((uint32_t)a[11*j+10] & 0xff) << 3)) * KYBER_Q) + 1024) >> 11);
     }
     a += 352;
   }
@@ -96,10 +96,10 @@ void PQCLEAN_NAMESPACE_polyvec_decompress(polyvec * restrict r, const uint8_t * 
   {
     for(j=0;j<KYBER_N/4;j++)
     {
-      r->vec[i].coeffs[4*j+0] =  (((a[5*j+ 0]       | (((uint32_t)a[5*j+ 1] & 0x03) << 8)) * KYBER_Q) + 512) >> 10;
-      r->vec[i].coeffs[4*j+1] = ((((a[5*j+ 1] >> 2) | (((uint32_t)a[5*j+ 2] & 0x0f) << 6)) * KYBER_Q) + 512) >> 10;
-      r->vec[i].coeffs[4*j+2] = ((((a[5*j+ 2] >> 4) | (((uint32_t)a[5*j+ 3] & 0x3f) << 4)) * KYBER_Q) + 512) >> 10;
-      r->vec[i].coeffs[4*j+3] = ((((a[5*j+ 3] >> 6) | (((uint32_t)a[5*j+ 4] & 0xff) << 2)) * KYBER_Q) + 512) >> 10;
+      r->vec[i].coeffs[4*j+0] = (int16_t)( (((a[5*j+ 0]       | (((uint32_t)a[5*j+ 1] & 0x03) << 8)) * KYBER_Q) + 512) >> 10);
+      r->vec[i].coeffs[4*j+1] = (int16_t)(((((a[5*j+ 1] >> 2) | (((uint32_t)a[5*j+ 2] & 0x0f) << 6)) * KYBER_Q) + 512) >> 10);
+      r->vec[i].coeffs[4*j+2] = (int16_t)(((((a[5*j+ 2] >> 4) | (((uint32_t)a[5*j+ 3] & 0x3f) << 4)) * KYBER_Q) + 512) >> 10);
+      r->vec[i].coeffs[4*j+3] = (int16_t)(((((a[5*j+ 3] >> 6) | (((uint32_t)a[5*j+ 4] & 0xff) << 2)) * KYBER_Q) + 512) >> 10);
     }
     a += 320;
   }
